Setup full round (still small bugs)
The mechanics ofrunning a fulll round are setup, from placing the cards, the betting round, and the revealing round. This is all done in the terminal from user inputs. There are still some notable bugs, which should be quick to fix at some point. Namely, when moving to the betting round the cursor of the player mooves forward. For example, if Player 0 ends the placing round the betting round will begin with Player 1 instead. There is also very little error checking. For example Players, when prompted to input 0, 1, or 2 can input 2 and end the placing round before everybody has placed a card, or enter an invalid number like 3 or -1. In the revealing round, there is nothing to stop you from trying to reveal a card from a player where all of their cards have already been revealed. This results in an error.