Simple game made
The game object is very simple and just runs a bunch of full rounds until there is a winner. A TestGame was created that can be run in the terminal to see that everything is working properly.These are not passing or failing tests though, just examined by playing the game at the moment.