Added error checking/misc bug fixes
Added error checking so that a player cannot break the game by inputting improper values. This was done for each 'piece' of the player by creating a set of possible valid integer inputs. So, each player piece now has a function validOptions() that gets the valid options. The RevealingRoundPlayer, currently handling two choices has a validOptions() function for flipping other players cards and removing your own in the case of failure. In hindsight, it is clear this doesn't align with the other 'pieces' of the player, and that it would make much more sense for each player to have a single validOptions() within it, and handle decisions of a single class. Despite these being tied closely when playing the game, it makes most sense to seperate them, so that will have to be done. Seperating them also makes sense for a deep learning model in the future, in which a neural network can handle each piece. Having validOptions() for each allows the network to choose from all possible options, ranking them, and then only performing the highest ranked option that is still valid. Additionally, this gives each network one responsibilty for each part of the game.