Added RemoveRoundPlayer
Created a player to perform the removal of the card after losing a bet. Since this is just a single player there was no need to make a round associated. This was put into the Player class as then used within runRound rather than performing this action through the RevealingRound and RevealingRoundPlayer. Some of this infrastructure could be cleaned for sure though. One more thing before the game is essentially complete (although messy) is removing players when they are out of the game. Naturally, this is a higher level above individual FullRounds or parts of a FullRound and will be handled by the Game object.