Game now removes players/handles winners
The game now properly handles ending the rounds when there is a winner, and removes players that have no more cards to play. However, at the moment there are two odd caveats. The first is that since the player name printed is fully dependent on the array of players passed, Player 0 (index 0) will change after a player is removed. For example, if there are three players, Player 0, Player 1, and Player 2, and Player 0 is removed, Player 2 will become Player 1 and Player 1 will become Player 0. This causes no actual issues with the Player objects nor does it affect the final print output of the players after the game is over, since that uses the list of all Player's instead of just active ones. This is still obviously very confusing, but could be fixed by changing these statements and adding names to the Player objects.

The second sort of odd thing is way the game is designed, so that when player is left, they win be default. This is a nice shortcut, but technically speaking, this player (if they didn't have a skull left) could just place their own card, and then bet on it since they are the only player left. This could be a good litmus test to see if the AI is working and can learn for a single player game by removing this small rule. The second thing, that I left unimplemented because the rule is not useful, is if a player reveals a skull as their final card they are own. Strategically, this player can't get anymore points and since the other players all know their final card, they should never be tricked with it. The only edge case where removing these players is useful is if there is a final player with only a skull and the other player gets out, so they win by default. I'm happy to have this small deviation from the typical ruleset.