Betting player bug fix
Playing the game had a bug in which after all of the players had placed their cards, the player who began the betting round was NOT the player forced to make the first bet. The cursor instead moved to the next player. This was an incredibly simple fix within the PlacingRound.