Players referred by names
Each of the players now have names. These are names that referred to in the revealing round now when asked which player to reveal a card from. This sidesteps the issue in which after a player was removed the index within the list changed. While there were other ways to do this without introducting names, names add little overhead and are nice when playing the game to have instead of just index numbers, even if the primary goal is to interface with an AI. This required each of the round classes to take players as their input instead of the specific type of player. I like this more anyway though since it removes the functions having to get the specific type of player from an object of type Player[], and it makes it a bit more homogenous. It also is nice since it allows the other 'sub player types' within player to be hidden, so just the broad player class can be used within exposing more of its internals.