Fixed game nuance bugs
While most of the game was working there were two bugs that deviating from the behvaior of the game, and thus the strategy somewhat. The first was MUCH more important than the second. Before this, the first player was always the first player in the game, and while the first player at the very beginning of the game has a slight advantage having every round start with them doesn't align with the game. In the game, the person who last bet begins the next placing round, so if Bob bet 3, whether he won or lost that bet for points, Bob will begin the next round. This was implemented. The second bug is much more bug like, and doesn't have an actual impact on the game (if you are playing as a human) but is really confusing. Before the bug fix the starting bet was 0. However, in the game, you must bet at least the number of cards that you have played. This makes sense, since to win a bet you must first reveal ALL of your cards. For example, if you have two cards, with a flower on top and a skull on the bottom, you must bet at least two, because if you win the bet, you must be obliged to reveal all of your own cards first. What makes this less of a gameplay issue is that the implementation of revealing your cards first knows this, and as such, even if you bet 1, it does not check that your bet is less than the number of cards played because that is assumed, so it would take 2 off, find that the bet is -1 after, and either give a point or prompt a card removal from there. The issue lies in what is allowed. It was still allowed to bet 1 when you had a stack of two, even though it didn't create a huge advantage (although a slight gameplay difference is that this allows the NEXT player to bet two, but I don't think that's significant).