Comment cleanup
Cleaned up some comments on potential future concrens or ideas, many of which were already resolved. The ones that were not a decision was made (usually to keep the current implementation). Next, the player classes need to become abstract classes with extending behavior for the console player and the AI, and wrapped more effectively in the Player class. Then, interfacing with a neuroevolution technique in python should be easier.