Document players / refactor placing decide
Added some documentation to each of parts for the "types" of players, such as the betting round player etc. The decide method for the placing player was returning a boolean of if to end the placing round while placing the card within the decide method. This doesn't make sense for what the decide method should do it now returns a integer representing the decision, and based on that decision the PlacingRound handles it. This of course makes it more flexible for the NEAT in the future, as decide can simply be hooked up to the neural net. Next, the NEAT will have to be developed to hook into the game.