Added NEAT Players
Added functional NEAT Players to the game. They can make decisions and play vis sockets with neural networks in Python. They don't learn anything yet though. Additionally, they keep track of their fitness based on how off their guesses were for valid moves. At the moment it is quite messy (lots of TODOs for documentation, and testing), and all of the cursory testing and debugging was just done in the main function within the new NEAT classes, although it seems to be working properly. The biggest issue is that after a player is eliminated, the neural network is given less inputs and breaks. So for now, two players is good. This could be fixed by allowing a node representing 0 cards in a player's hand, indicating they are no longer in the game.

There was one bug fix to the base game:
When a player is eliminated, the cursor wasn't updated, which works to shift the player forward, unless the cursor is placed on the final player, in which case it should loop around to zero, being otherwise out of range. This edge case was added and works now.