Set number of games in population
When running games throughout a population, it can now play random games with players, not repeating those games. A BUG is that with >2 players it seems to hang. I think this may be a bug in the game that happens very rarely and only with >2 players, causing an infinite loop in the game logic on the java end.