Can Player Trained NEAT networks
NEAT Networks can now be trained, and using the NEAT checkpoint and the correct genome on run_specific_network_socket to open the socket, and then running Game with that socket on a NEATPlayerVerbose (allows the action of the game to be seen in the terminal) a basic neural network trained by neat can be played against. This is NOT a recurrent neural network and has no memory. Softmax is not also handled on the output of the neural network for rankings. Still quite messy, and could use refactoring/cleaning..